style(missing files): update pages imports for file renaming
